implement test_select_dtypes (#32250)
